Support "positive" filtering
Previously, the filtering mechanism was “hide all messages which involve
one of these names”, although the dialog presented them in the opposite
sense. I often want to say something a bit different: “show only
messages which involve one of these names”.

Implement this with three states for each name: Default, Only, and
Never. The logic for each message is:

 * If it involves at least one Never name, hide it
 * If there are no Only names, show it
 * Otherwise, show it if it involves at least one Only name