icons: switch SVGs from 90 DPI to 96 DPI
When opening these icons, Inkscape reports:

    $FILENAME was created in an older version of Inkscape (90 DPI) and
    we need to make it compatible with newer versions (96 DPI).

with a link to https://inkscape.org/en/learn/faq#dpi_change for more
details.

These are intended for screen display, so I picked the option to not
rescale at all; in practice this did cause the XML to be rewritten in
mostly-harmless ways.