Reimplement core functionality in C
Obviously I did not write this all in one go, but I didn't preserve my
incremental work. bustle-window.ui is almost unmodified from
data/bustle.ui, and the build system builds the existing C code, but
otherwise it's all new.

The key part is that the log is represented as a GtkTreeModel
(specifically a GtkListStore) with references between method calls and
the corresponding reply¹ are represented with a GtkTreeRowReference. The
chart is drawn using a GtkTreeView and a custom renderer that renders
one horizontal stripe of the chart at a time. The approach (and the
colour scheme for the lanes) is inspired by gitg. The current
name-ownership state is tracked explicitly for each row of the diagram
(using an externally-immutable data structure, naturally).

By using the GtkTreeView machinery, we get many things essentially for
free that I've failed to implement in the old codebase for over a
decade, including type-ahead search (30 minutes) and selection tracking
(surely less buggy than my implementation).

There are many missing features:

- None of the UI is wired up outside the diagram itself and the details
  pane. This is just a matter of typing.
- The rendering of each message is incomplete: undirected signals are
  simply not drawn, for example.
- Support for showing system and session bus logs side-by-side. I think
  this will be relatively straightforward: it's just™ a matter of a few
  extra model columns and another TreeViewColumn.

I have not worked out how to fit some aspects of the old diagram into
the new world order:

- Arcs between method calls and the reply are really hard to do when you
  fundamentally don't know where the other side will be. I was always
  very unsatisfied with how this looked for very busy diagrams in the
  old codebase so maybe I can find another way to display this
  anyway.
- Headers for each "lane" in the diagram, showing which name(s) it owns.
  Again, I was never satisfied with how it looked before (taking only
  the last component of each name worked fine for Telepathy but is
  useless in many other cases) so I think this will just stay
  unimplemented. I'll either add a colour-coded legend, or something to
  the details pane, or both.

¹ or replies: the edge case where more than one reply is seen for a
  given message is handled