Don't share owned_names values across copy()
Previously, owned_names itself was copied, but the GHashTables which are
its values were not. This means that name ownership changes would affect
previous instances in the chain.

As explained in the TODO, we could do this more lazily for a
potentially-large memory saving. (In fact, I thought I had done so
already, but no.) Let's do it the memory-hungry and slow but safe way
for now.