pcap-monitor: clear G_MESSAGES_DEBUG from environment
g_log_default_handler() is documented to write INFO and DEBUG messages
to stdout. flatpak-spawn emits such messages if run with
G_MESSAGES_DEBUG=all, and its stdout is by necessity shared with its
child.

In this case, the child is dbus-monitor. Running Bustle with
G_MESSAGES_DEBUG=all in a Flatpak sandbox, then trying to record a log,
would cause the pcap-format message stream to be interleaved with output
like:

    ** (flatpak-spawn:116): DEBUG: 21:54:39.917: child_pid: 25091

and libpcap complains that 2a 2a 20 28 is not a known format header.